sequential_track_playback-manual.html

<!doctype html>
<html>

<head>
  <meta charset=utf-8>
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
  <video style="height: 400px"></video>
  <div>Press Enter/Ok to continue or Up/Down to show/hide debug overlay</div>
  <div id="debug"></div>
  <div id=log></div>
  <div id="info-overlay"></div>
</body>
<script src="lib/mozilla/object-keys-polyfill.js"></script>
<script src="lib/stefanpenner/es6-promise.min.js"></script>
<script src="lib/utils/test_helper.js"></script>
<script src="lib/player.js"></script>
<script src="lib/manifest_parser.js"></script>
<script src="lib/mpd-parser.js"></script>
<script src="sub-tests/every_sample_should_be_rendered.js"></script>
<script src="sub-tests/playback_duration_matches_cmaf_duration.js"></script>
<script src="sub-tests/low_startup_delay.js"></script>
<script src="lib/info-overlay.js"></script>
<script>
  var TEST_INFO = {
    title: "Sequential track playback",
    description: "Sequential track playback refers to the case that a CMAF/WAVE track is played from the beginning by providing CMAF fragments to the source buffer after initialization."
  };
  document.getElementsByTagName("title")[0].innerText = TEST_INFO.title;

  setup({
    explicit_timeout: true
  });
  var video = document.querySelector('video');
  player = new Player(video);
  player.init({bufferTime: 30, numberOfSegmentBeforePlay: 2});

  var infoOverlayElement = document.getElementById("info-overlay");
  var infoOverlay = new InfoOverlay(infoOverlayElement);
  infoOverlay.init();

  function updateOverlayInfo() {
    var manifest = player.getManifest();
    var playingRepresentation = player.getPlayingRepresentation();
    var number = "none";
    if (playingRepresentation) number = playingRepresentation.getNumber();
    infoOverlay.setInfo({
      test: TEST_INFO,
      video: {
        representations: manifest.getRepresentations(),
        playingRepresentation: number,
        currentTime: player.getCurrentTime(),
        playingSegment: player.getPlayingSegment().getNumber(),
        duration: manifest.getDuration(),
      }
    });
    infoOverlay.renderInfo();
  }

  //every_sample_shall_be_rendered();
  //playback_duration_matches_cmaf_duration();
  low_startup_delay(120, player.settings.numberOfSegmentBeforePlay);

  var contentModel = urlParams["content_model"];
  player.load(contentModel);
  player.addEventListener("onManifestParsed", function (manifest) {
    var totalSegmentsCount = manifest
      .getRepresentation(1)
      .getTotalSegmentsCount();
    player.setVideoSegments({
      representationNumber: 1,
      startSegment: 10,
      endSegment: totalSegmentsCount - 1,
    });
    player.startBuffering();
    updateOverlayInfo();
  });
  player.addEventListener("onSegmentLoaded", function (event) {
    if (event.totalSegmentsLoaded === 2) {
      player.play();
    }
  });
  player.addEventListener("onPlayingRepresentationChange", function (currentRepresentation) {
    updateOverlayInfo();
  });
  player.addEventListener("onPlayingSegmentChange", function (segment) {
    //console.log(segment);
  });
  player.addEventListener("onTimeUpdate", function (currentTime) {
    updateOverlayInfo();
  });

  async_test(function (test) {
    document.addEventListener("keydown", test.step_func(function (event) {
      assert_true(event.keyCode === 13);
      test.done();
    }));
  }, "Placeholder subtest: Press Enter");

  document.addEventListener("keydown", function (event) {
    if (event.keyCode === 38) {
      infoOverlay.show();
    }
    if (event.keyCode === 40) {
      infoOverlay.hide();
    }
  });
</script>
</body>

</html>
